:recycle: refactor: validation 분리
