+++ src/main/java/gift/util/validator/entityValidator/MemberValidator.java
@@ -1,4 +1,4 @@
-package gift.validator.entityValidator;
+package gift.util.validator.entityValidator;
 
 import gift.dto.MemberDTO;
 import java.util.regex.Pattern;
@@ -9,18 +9,18 @@ public class MemberValidator {
             "^[A-Za-z0-9+_.-]+@(.+)$"
     );
 
-    public void validate(MemberDTO memberDTO) {
+    public static void validate(MemberDTO memberDTO) {
         validateEmail(memberDTO.getEmail());
         validatePassword(memberDTO.getPassword());
     }
 
-    private void validateEmail(String email) {
+    private static void validateEmail(String email) {
         if (email == null || !EMAIL_PATTERN.matcher(email).matches()) {
             throw new IllegalArgumentException("올바른 이메일 형식이 아닙니다.");
         }
     }
 
-    private void validatePassword(String password) {
+    private static void validatePassword(String password) {
         if (password == null || password.isBlank()) {
             throw new IllegalArgumentException("비밀번호를 입력해주세요.");
         }
